+++ apps/server/src/entities/entities.controller.ts
@@ -12,7 +12,7 @@ import { EntitiesService } from './entities.service'
 import { NotFoundInterceptor } from '../middlewares/not-found.interceptor'
 
 import { Actor, Entity } from './entities.interface'
-import { HEADER_GLOBAL_STORAGE_INSTANCE_ID, HEADER_GLOBAL_STORAGE_USER_ID } from 'global-storage'
+import { HEADER_GLOBAL_STORAGE_INSTANCE_ID, HEADER_GLOBAL_STORAGE_USER_ID } from 'central-storage'
 
 const publicApiPrefix = '/entities/'