+++ packages/js-client/src/core/constants.ts
@@ -1,2 +1,2 @@
-export const HEADER_GLOBAL_STORAGE_INSTANCE_ID = 'x-central-storage-instance-id'
-export const HEADER_GLOBAL_STORAGE_USER_ID = 'x-central-storage-user-id'
+export const HEADER_GLOBAL_STORAGE_INSTANCE_ID = 'x-remote-storage-instance-id'
+export const HEADER_GLOBAL_STORAGE_USER_ID = 'x-remote-storage-user-id'